+++ views/login.handlebars
@@ -18,4 +18,4 @@
 
     </div>
 </div>
-<script src="/assets/js/auth.js"></script>
+<script src="/assets/js/auth-r.js"></script>
